# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: 3D Marketplace API
  version: "0.1.0"
  description: AR-powered 3D product marketplace API
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/products/publish:
    post:
      operationId: publishProduct
      summary: Publish a READY asset as a product
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "201":
          description: Product published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/products:
    get:
      operationId: listProducts
      summary: List published products
      tags: [products]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: seller_id
          in: query
          description: Filter products by seller UUID
          schema:
            type: string
            format: uuid
        - name: liked
          in: query
          description: "If true, return only products liked by the authenticated user (requires auth)"
          schema:
            type: boolean
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /v1/products/{id}:
    get:
      operationId: getProduct
      summary: Get product detail
      tags: [products]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/products/{id}/like:
    post:
      operationId: toggleProductLike
      summary: Toggle like on a product
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeToggleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/products/{id}/purchase:
    post:
      operationId: purchaseProduct
      summary: Purchase a product (creates purchase record and marks product SOLD_OUT)
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Purchase created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/products/{id}/ar-asset:
    get:
      operationId: getProductArAsset
      summary: Get AR asset for a product
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: AR asset info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArAssetResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/products/{id}/chat-rooms:
    post:
      operationId: createProductChatRoom
      summary: Create a chat room for a product
      tags: [chat]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatRoomRequest"
      responses:
        "201":
          description: Chat room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatRoomResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
      required:
        - error

    # --- Upload ---
...
    ArAssetResponse:
      type: object
      properties:
        availability:
          type: string
          enum: [READY, PROCESSING, NONE]
        asset_id:
          type: string
          format: uuid
          nullable: true
        files:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              url:
                type: string
                format: uri
              type:
                type: string
                enum: [model, preview]
            required:
              - role
              - url
              - type
        dims_source:
          type: string
          enum: [ios_lidar, ios_manual, unknown]
          nullable: true
        dims_trust:
          type: string
          enum: [high, medium, low]
          nullable: true
      required:
        - availability
        - files

    # --- Product ---
    PublishRequest:
      type: object
      properties:
        asset_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price_cents:
          type: integer
      required:
        - asset_id
        - title
        - price_cents
    ProductStatus:
      type: string
      enum: [FOR_SALE, RESERVED, SOLD_OUT]
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        price_cents:
          type: integer
        seller_id:
          type: string
          format: uuid
        published_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        seller_name:
          type: string
          default: ""
        seller_avatar_url:
          type: string
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ProductStatus"
        likes_count:
          type: integer
          default: 0
        views_count:
          type: integer
          default: 0
        chat_count:
          type: integer
          default: 0
        seller_location_name:
          type: string
          nullable: true
        is_liked:
          type: boolean
          nullable: true
      required:
        - id
        - asset_id
        - title
        - price_cents
        - seller_id
        - created_at
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - products
        - total
        - page
        - limit
    PurchaseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        buyer_id:
          type: string
          format: uuid
        price_cents:
          type: integer
        purchased_at:
          type: string
          format: date-time
        product:
          $ref: "#/components/schemas/ProductResponse"
      required:
        - id
        - product_id
        - buyer_id
        - price_cents
        - purchased_at
...
    LikeToggleResponse:
      type: object
      properties:
        liked:
          type: boolean
        likes_count:
          type: integer
      required:
        - liked
        - likes_count

    # --- Chat ---
    CreateChatRoomRequest:
      type: object
      properties:
        subject:
          type: string
      required:
        - subject
    ChatRoomResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        buyer_id:
          type: string
          format: uuid
        seller_id:
          type: string
          format: uuid
        subject:
          type: string
        created_at:
          type: string
          format: date-time
        last_message_at:
          type: string
          format: date-time
          nullable: true
        last_message_body:
          type: string
          nullable: true
        unread_count:
          type: integer
          default: 0
        buyer_name:
          type: string
          default: ""
        seller_name:
          type: string
          default: ""
        product_title:
          type: string
          default: ""
        product_thumbnail_url:
          type: string
          nullable: true
      required:
        - id
        - product_id
        - buyer_id
        - seller_id
        - subject
        - created_at
...